[15.0][MIG] mail_activity_done: Migration to 15.0
