+++ mail_activity_done/readme/USAGE.rst
@@ -1,0 +1,9 @@
+To use this module, you need to:
+
+#. Access to an activity from the systray activities menu.
+#. Once finished, open the activity and mark it as Done.
+
+To check activities:
+
+#. Go to Settings -> Technical -> Activities
+#. To see finished activities filter by Completed Activities
