+++ mail_activity_done/tests/test_mail_activity_done.py
@@ -35,7 +35,7 @@ class TestMailActivityDoneMethods(TransactionCase):
         self.assertEqual(self.act1.state, "done")
 
     def test_systray_get_activities(self):
-        act_count = self.employee.sudo(self.employee).systray_get_activities()
+        act_count = self.employee.with_user(self.employee).systray_get_activities()
         self.assertEqual(
             len(act_count), 1, "Number of activities should be equal to one"
         )
