--- mail_activity_done/views/templates.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <!--
-        Copyright 2018-20 ForgeFlow <http://www.forgeflow.com>
-        License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
-    -->
-    <template
-        id="assets_backend"
-        name="mail_activity_done assets"
-        inherit_id="web.assets_backend"
-    >
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/mail_activity_done/static/src/js/mail_activity.js"
-            />
-        </xpath>
-    </template>
-</odoo>
